% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.cases.r
\name{add.cases}
\alias{add.cases}
\title{FUNCIO QUE AFEGEIX CASOS
funcio que afegeix els casos de y sota dels de x.}
\usage{
add.cases(x, y, font = NULL, all = 3, show.warnings = TRUE)
}
\arguments{
\item{x}{primera base de dades}

\item{y}{segona base de dades que es posara a sota}

\item{font}{NULL}

\item{all}{1: posa les variables comunes de x i de y; 2: posa totes les variables de x; 3: posa totes les variables de x i totes les variables de y.}
}
\description{
FUNCIO QUE AFEGEIX CASOS
funcio que afegeix els casos de y sota dels de x.
}
\note{
per a l'opcio all=1 i all=2 manen les etiquetes i formats de x
per a l'opcio all=3 manen les etiquetes de x per a les variables comunes i les de y per a la resta. Els formats resultants s?n
coherents amb la fusio (realitza un arregla.formats)
}
