% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arregla.formats.r
\name{arregla.formats}
\alias{arregla.formats}
\title{arregla formats}
\usage{
arregla.formats(taula, identif = NULL, force = FALSE,
  rate.miss.date = 0.5)
}
\arguments{
\item{taula}{data.frame o matrix}

\item{identif}{NULL}

\item{rate.miss.date}{= 0.5}

\item{force:}{si és TRUE arregla els formats, encara que l'atribut "formats.arreglats" sigui TRUE.}
}
\description{
arregla formats
}
\note{
si la variable data no està definida com a chron, aleshores hauran d'està en el següent format, si no ho deixa a caràcter.
les dates estan separades per algun d'aquests caràcters: '-', '/', '.'.
les hores estan separades per el caràcter ':'.

si la variable està en format data, o sigui, chron, estarà en el format (veure ?chron):
les dates day-mon-years
les hores com a h:m:s
}
