% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arregla.miss.values.r
\name{arregla.miss.values}
\alias{arregla.miss.values}
\title{arregla.miss.values
reemplaça els miss.values de una variable per NA, i assigna TRUE a un nou atribut anomenat miss.values.arreglat.}
\usage{
arregla.miss.values(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
arregla.miss.values
reemplaça els miss.values de una variable per NA, i assigna TRUE a un nou atribut anomenat miss.values.arreglat.
}
