% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.fresco.r
\name{calculadora.fresco.curt}
\alias{calculadora.fresco.curt}
\title{calculadora.fresco}
\usage{
calculadora.fresco.curt(sexo, edad, fuma, imc, rang.edat = c(35, 79),
  anys = 10)
}
\description{
calculadora.fresco
}
