% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.risc.gen.r
\name{calculadora.risc.gen}
\alias{calculadora.risc.gen}
\title{calculadora.risc.gen
Calculadora risc REGICOR/FRAMINGHAM que té en compte la genètica}
\usage{
calculadora.risc.gen(sex, age, coltot, hdl, tas, tad, diab, smoke, snps,
  comp.gen, calibrated = TRUE, age.range = c(35, 74), betas.gen.men,
  betas.gen.wom)
}
\arguments{
\item{sex}{1:hombres, 2:mujer ( o otro n?mero diferente de zero)}

\item{age}{edad en a?os}

\item{coltot}{colesterol total en mg/dL}

\item{hdl}{colesterol HDL en mg/dL}

\item{tas}{TAS en mmHg}

\item{tad}{TAD en mmHg}

\item{diab}{diabetes (1=s?, 0=no)}

\item{smoke}{fumador actual o <1a (1=s?, 0=no)}

\item{snps}{matrix on cada fila ?s el nombre d'alels de risc (0, 1 ? 2) per a cada snp, per a cada individu.}

\item{comp.gen}{-0.2462213 # suma del productes de les betes per les prevalences dels alels de risc.}

\item{calibrated}{calibrada (TRUE) o original (FALSE).}

\item{betas.gen.men}{vector betas SNPs en homes}

\item{betas.gen.men}{vector betas SNPs en dones}
}
\description{
calculadora.risc.gen
Calculadora risc REGICOR/FRAMINGHAM que té en compte la genètica
}
