% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.risc.hard.r
\name{calculadora.risc.hard}
\alias{calculadora.risc.hard}
\title{calculadora.risc.hard
Calculo del riesgo en enfermedad coronaria con las funciones de Framingham  y la adaptada de REGICOR}
\usage{
calculadora.risc.hard(sex, age, coltot, hdl, tas, tad, diab, smoke,
  calibrated = TRUE, age.range = c(35, 74), g_chol.men = 4.2935939,
  g_chol.wom = 6.5843, S0.men = 1 - 0.035, S0.wom = 1 - 0.011)
}
\arguments{
\item{sex}{1:hombres, 2:mujer ( o otro n?mero diferente de zero)}

\item{age}{edad en a?os}

\item{coltot}{colesterol total en mg/dL}

\item{hdl}{colesterol HDL en mg/dL}

\item{tas}{TAS en mmHg}

\item{tad}{TAD en mmHg}

\item{diab}{diabetes (1=s?, 0=no)}

\item{smoke}{fumador actual o <1a (1=s?, 0=no)}

\item{calibrated}{funci? calibrada (TRUE) o original (FALSE).}

\item{g_chol.men, }{g_chol.wom:  sum(betes*mean(x))}

\item{S0.men}{supervivencia basal en hombres (en la media de las covariables)}

\item{S0.wom:}{supervivencia basal en mujeres (en la media de las covariables)}
}
\description{
calculadora.risc.hard
Calculo del riesgo en enfermedad coronaria con las funciones de Framingham  y la adaptada de REGICOR
}
\examples{
calculadora.risc.hard(sex=1,
                     age=55,
                     coltot=154,
                     hdl=46,
                     tas=110,
                     tad=60,
                     diab=1,
                     smoke=1,
                     calibrated=TRUE,age.range=c(35,74))
}
