% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.risc.r
\name{calculadora.risc}
\alias{calculadora.risc}
\title{calculadora.risc
Calculo del riesgo en enfermedad coronaria con las funciones de Framingham  y la adaptada de REGICOR}
\usage{
calculadora.risc(sex, age, coltot, hdl, tas, tad, diab, smoke,
  calibrated = TRUE, age.range = c(35, 74))
}
\arguments{
\item{sex}{1:hombres, 2:mujer (o cualquier n?mero diferente de uno)}

\item{age}{edad en a?os}

\item{coltot}{colesterol total en mg/dL}

\item{hdl}{colesterol HDL en mg/dL}

\item{tas}{TAS en mmHg}

\item{tad}{TAD en mmHg}

\item{diab}{diabetes (1=s?, 0=no)}

\item{smoke}{fumador actual o <1a (1=s?, 0=no)}

\item{calibrated}{funci? calibrada (TRUE) o original (FALSE).}
}
\description{
calculadora.risc
Calculo del riesgo en enfermedad coronaria con las funciones de Framingham  y la adaptada de REGICOR
}
