% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.risc5a.r
\name{calculadora.risc5a}
\alias{calculadora.risc5a}
\title{calculadora.risc5a
calculadora risc REGICOR a 5 anys}
\usage{
calculadora.risc5a(sex, age, coltot, hdl, tas, tad, diab, smoke,
  calibrated = TRUE, age.range = c(35, 74))
}
\arguments{
\item{sex}{1:hombres, 2:mujer ( o otro n?mero diferente de zero)}

\item{age}{edad en a?os}

\item{coltot}{colesterol total en mg/dL}

\item{hdl}{colesterol HDL en mg/dL}

\item{tas}{TAS en mmHg}

\item{tad}{TAD en mmHg}

\item{diab}{diabetes (1=s?, 0=no)}

\item{smoke}{fumador actual o <1a (1=s?, 0=no)}

\item{calibrated}{funci? calibrada (TRUE) o original (FALSE).}
}
\description{
calculadora.risc5a
calculadora risc REGICOR a 5 anys
}
