% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.risc.gen.r, R/calculadora.score.r
\name{calculadora.score}
\alias{calculadora.score}
\title{calculadora.score
Calculadora risc SCORE que té en compte la genètica}
\usage{
calculadora.score(sexo, edat, coltot, tas, fuma2, diabetes, anys = 5,
  low.risk = TRUE)

calculadora.score(sexo, edat, coltot, tas, fuma2, diabetes, anys = 5,
  low.risk = TRUE)
}
\arguments{
\item{sexo}{(0 dones, 1 homes)}

\item{edat}{edat}

\item{coltot}{(mg/dl)}

\item{tas}{(mmHg)}

\item{fuma2}{(0 mai ? ex>1a; 1 - actual ? ex<1a)}

\item{diabetes}{(0 no, 1 s?)}

\item{sexo}{(0 dones, 1 homes)}

\item{edat:}{ha de ser >=45 i <=64}

\item{coltot}{(mg/dl)}

\item{tas}{(mmHg)}

\item{fuma2}{(0 mai ? ex>1a; 1 - actual ? ex<1a)}

\item{diabetes}{(0 no, 1 s?)}
}
\description{
calculadora.score
Calculadora risc SCORE que té en compte la genètica

calculadora.score
CALCUL RISC SCORE
}
