% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega.llibreria.r
\name{carrega.llibreria}
\alias{carrega.llibreria}
\title{carrega.llibreria
fa un library i un install.packages (si cal)}
\usage{
carrega.llibreria(llibreria)
}
\arguments{
\item{llibreria:}{vector de character amb el nom de les llibreries a descarregar.}
}
\description{
carrega.llibreria
fa un library i un install.packages (si cal)
}
