% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causa.exitus_v2.r
\name{causa.exitus_v2}
\alias{causa.exitus_v2}
\title{causa.exitus_v2
causa exitus}
\usage{
causa.exitus_v2(cie, status = 9, fexit = NULL, cexit = NULL,
  sexitus = 1, sobran = 0)
}
\arguments{
\item{cie}{cie9 or cie10}

\item{status}{Por defecto "1='muerto'"}

\item{fexit}{'diaexit' or 'mesexit' or 'anyoexit'}

\item{cexit}{default NULL}

\item{sobran}{si es '0', se devuelve la causa de exitus, si es '>0' se devuelve 'xcieIsobran' (si es CIE que sobra o no) (los CIEs que sobran)}

\item{sexitus<-Aqui}{se pondra el valor que tienen los muertos en la variable 'status'}
}
\description{
causa.exitus_v2
causa exitus
}
