% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idi_surv.R
\name{ciidi}
\alias{ciidi}
\title{ciidi
Calcula l'IDI (índex de reclassificació) amb interval de confiança}
\usage{
ciidi(dat, rr1, rr2, outcome, B = 1000, alpha = 0.05, seed,
  digits = 2)
}
\arguments{
\item{dat}{dades}

\item{rr1}{risc segons el model 'vell'}

\item{rr2}{risc segons el model 'nou'}

\item{outcome}{nom de la variable cas-control (binaria)}

\item{B}{nombre d'iteracions (default 1000)}

\item{alpha}{nivell de significació (default 0.05)}

\item{seed}{llavor}

\item{digits}{digits (default 2)}
}
\description{
ciidi
Calcula l'IDI (índex de reclassificació) amb interval de confiança
}
