% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.r
\name{Classification}
\alias{Classification}
\title{Classification
Retorna la freq. de les combinacions de variables i la categoria a la que correspon cada individu
Mostra els NA com a 9999  (selec=2 fa que retorni la categoria a cada individu)}
\usage{
Classification(dades, vars, selec = 1)
}
\arguments{
\item{dades}{base de dades}

\item{vars}{vector amb els noms de les variables}

\item{selec}{(default is 1)}
}
\description{
Classification
Retorna la freq. de les combinacions de variables i la categoria a la que correspon cada individu
Mostra els NA com a 9999  (selec=2 fa que retorni la categoria a cada individu)
}
