% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.qualitat.r
\name{control.qualitat}
\alias{control.qualitat}
\title{control.qualitat
genera un informe amb estadístics per a fer control de qualitat}
\usage{
control.qualitat(file = NULL, data.intern = FALSE, max.valors = 10,
  desc = NULL, max.rare.values = 4, max.factor = max.valors,
  digits = 1, file.res = "xxx.xls")
}
\arguments{
\item{file}{NULL}

\item{data.intern}{FALSE}

\item{max.valors}{10}

\item{desc}{NULL}

\item{max.rare.values}{4}

\item{max.factor}{max.valors}

\item{digits}{1}

\item{file.res}{"xxx.xls"}
}
\description{
control.qualitat
genera un informe amb estadístics per a fer control de qualitat
}
