% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.dates.r
\name{convert.dates}
\alias{convert.dates}
\title{convert.dates
crea una variable de tipus data (chron) a partir del dia, mes i any (integers)}
\usage{
convert.dates(day, month, year)
}
\arguments{
\item{day}{dies (vector integer)}

\item{month}{mes (vector integer)}

\item{year}{any (vector integer)}
}
\description{
convert.dates
crea una variable de tipus data (chron) a partir del dia, mes i any (integers)
}
