% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crea.haplo.r
\name{crea.haplo}
\alias{crea.haplo}
\title{crea.haplo
Crea haplotips a partir de SNPs}
\usage{
crea.haplo(data, locus.labels, haplo.allele)
}
\arguments{
\item{locus.labels}{noms dels SNPs}

\item{dades}{que inclouen els SNPs}

\item{allel}{de l'haplotip}
}
\description{
crea.haplo
Crea haplotips a partir de SNPs
}
