% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crea.haplo2.r
\name{crea.haplo2}
\alias{crea.haplo2}
\title{crea.haplo2
Crea haplotips a partir de SNPs tenint en compte els pesos}
\usage{
crea.haplo2(data, locus.labels, haplo.allele, pesos)
}
\arguments{
\item{locus.labels}{noms dels SNPs}

\item{pesos}{pesos dels SNPs}

\item{dades}{que inclouen els SNPs}

\item{allel}{de l'haplotip}
}
\description{
crea.haplo2
Crea haplotips a partir de SNPs tenint en compte els pesos
}
