% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut3.r
\name{cut3}
\alias{cut3}
\title{cut3
fa quantils amb el mateix mètode que ho fa el SPSS}
\usage{
cut3(x, g)
}
\arguments{
\item{x}{variable numèrica}

\item{g}{nombre de quantils (amb la mateixa n)}
}
\description{
cut3
fa quantils amb el mateix mètode que ho fa el SPSS
}
