% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaAUC_surv.r
\name{deltaAUC.surv}
\alias{deltaAUC.surv}
\title{deltaAUC.surv
funció que calcula la diferència de AUC (C'index) entre dos models predictius tenint en compte la supervivència}
\usage{
deltaAUC.surv(dat, ref, new, times, outcome)
}
\arguments{
\item{dat}{dades}

\item{ref}{predits segons el model 'vell'}

\item{new}{predits segons el model 'nou'}

\item{times}{nom de la variable temps fins a event o censura}

\item{outcome}{nom de la variable indicadora de censura}
}
\description{
deltaAUC.surv
funció que calcula la diferència de AUC (C'index) entre dos models predictius tenint en compte la supervivència
}
