% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.ACCESS.r
\name{dict.ACCESS}
\alias{dict.ACCESS}
\title{dict.ACCESS
crea un fitxer de diccionari de variables (millor txt) per a importar a ACCESS d'uns fitxers de sps (en el mateix format del que dona la funci? dictionary.r)}
\usage{
dict.ACCESS(files.sps, file.res, nom.taules = gsub(".sps$", "",
  basename(files.sps)))
}
\arguments{
\item{files.sps}{fitxer sintaxis SPSS}

\item{file.res}{nom del fitxer de resultats}

\item{nom.taules}{nom de les taules}
}
\description{
dict.ACCESS
crea un fitxer de diccionari de variables (millor txt) per a importar a ACCESS d'uns fitxers de sps (en el mateix format del que dona la funci? dictionary.r)
}
