% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict2data.r
\name{dict2data}
\alias{dict2data}
\title{dict2data
incorpora un fitxer .sps de diccionari de variables ('file.dict') al data.frame 'dades' ###
i retorna dades etiquetat (vari.labels i value.labels) ########}
\usage{
dict2data(file.dict, dades)
}
\arguments{
\item{file.dict}{nom del fitxer del diccionari}

\item{dades}{dades}
}
\description{
dict2data
incorpora un fitxer .sps de diccionari de variables ('file.dict') al data.frame 'dades' ###
i retorna dades etiquetat (vari.labels i value.labels) ########
}
