% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.r
\name{dictionary}
\alias{dictionary}
\title{dictionary
CREA UN FITXER *.SPS DE VALUE LABELS I VARI LABELS D'UNA BASE DE DADES *.SAV ####}
\usage{
dictionary(files = NULL, file.res = "")
}
\arguments{
\item{files}{nom dels fitxers}

\item{file.res}{nom del fitxer resultat}
}
\description{
dictionary
CREA UN FITXER *.SPS DE VALUE LABELS I VARI LABELS D'UNA BASE DE DADES *.SAV ####
}
