% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.ACCESS.r
\name{export.ACCESS}
\alias{export.ACCESS}
\title{export.ACCESS
FUNCIO QUE EXPORTA UNA TAULA (data.frame) A UNA BASE DE DADES EXISTENT! DE ACCESS. A MES, POSA LES VARIABLES DATA EN FORMAT DATA!!!
TAMBE (OPCIONAL) CREA UNA TAULA AMB EL DICCIONARI (VARI LABELS, VALUE LABELS) DE LES VARIABLES DE TAULA.}
\usage{
export.ACCESS(taula, file.mdb, table.name, table.dict = NULL,
  fix.formats = TRUE)
}
\arguments{
\item{taula}{data.frame amb les dades per a exportar a ACCESS}

\item{file.mdb}{character, especifica el path i el filename de la base de dades EXISTENT! de ACCESS on va a parar 'taula'}

\item{table.name}{nom de la taula en el ACCESS}

\item{table.dict}{NULL - no crea cap diccionari, !NULL - crea una taula que sera el diccionari de 'taula' amb el nom especificat com a caracter.}

\item{fix.formats}{arreglar formats (default TRUE)}
}
\description{
export.ACCESS
FUNCIO QUE EXPORTA UNA TAULA (data.frame) A UNA BASE DE DADES EXISTENT! DE ACCESS. A MES, POSA LES VARIABLES DATA EN FORMAT DATA!!!
TAMBE (OPCIONAL) CREA UNA TAULA AMB EL DICCIONARI (VARI LABELS, VALUE LABELS) DE LES VARIABLES DE TAULA.
}
\note{
PER A MILLORAR: DE MOMENT NO ARREGLA LES VARIABLES EN FORMAT DATA AMB HORES:MINUTS:SEGONS (EN ACCESS QUEDA COM UN REAL PROCEDENT DEL FORMAT DE R).
}
