% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.SPSS.r
\name{export.SPSS}
\alias{export.SPSS}
\title{export.SPSS
EXPORTA TAULA (DATA.FRAME O MATRIX) DE R A SPSS}
\usage{
export.SPSS(taula, file.dict = NULL, file.save = NULL,
  var.keep = "ALL",
  file.runsyntax = "C:\\\\Archivos de programa\\\\SPSS\\\\runsyntx.exe",
  file.dades = NULL, run.spss = TRUE, dec = ".")
}
\arguments{
\item{taula}{base de dades}

\item{file.dict}{nom del fitxer diccionari}

\item{file.save}{nom de la base de dades guardada en SPSS}

\item{var.keep}{vector amb el nom de les variables a guardar ('ALL' per guardar-les totes)}

\item{file.runsyntax}{on està l'executable de SPSS (ó PSPP) per a guardar el fitxer .sav}

\item{file.dades}{nom del fitxer on es guarden les dades en format txt (default NULL - temporal)}

\item{run.spss}{executa l'SPSS (ó PSPP) per a generar el .sav (default TRUE)}

\item{dec}{caràcter decimal}
}
\description{
export.SPSS
EXPORTA TAULA (DATA.FRAME O MATRIX) DE R A SPSS
}
