% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix2.r
\name{fix2}
\alias{fix2}
\title{fix2
FUNCI? QUE VISUALITZA (a traves de la funcio 'fix') UNA MATRIU O UN DATA.FRAME #############}
\usage{
fix2(data, view = 1, print = FALSE, rows = NULL)
}
\arguments{
\item{data}{dades}

\item{view}{=1: noms etiquetes, view=2: valor i etiqueta, view!=1  2: noms valor}

\item{print}{TRUE: mostra el data.frame per la consola de R, print=FALSE: mostra el data.frame per una finestra (com la funcio 'fix2')}

\item{rows:}{selecciona els individus i ha de ser el nombre de la fila de l'individu en el data.frame.}
}
\description{
fix2
FUNCI? QUE VISUALITZA (a traves de la funcio 'fix') UNA MATRIU O UN DATA.FRAME #############
}
