% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format2.r
\name{format2}
\alias{format2}
\title{format2
similar a format però posa el digits per defecte de forma 'intel.ligent'.}
\usage{
format2(x, digits = NULL, ...)
}
\arguments{
\item{x}{vector numèric}

\item{digits}{nombre de decimals}
}
\description{
format2
similar a format però posa el digits per defecte de forma 'intel.ligent'.
}
