% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabels.r
\name{getLabels}
\alias{getLabels}
\title{getLabels
recull les etiquetes de variable i de valors d'una base de dades}
\usage{
getLabels(data)
}
\arguments{
\item{data}{base de dades}
}
\description{
getLabels
recull les etiquetes de variable i de valors d'una base de dades
}
