% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HosmerLemeshow.r
\name{HosmerLemeshow.coxph}
\alias{HosmerLemeshow.coxph}
\title{HosmlerLemeshow.coxph
Realitza el test de Hosmer-Lemeshow a partir d'un model de Cox}
\usage{
\method{HosmerLemeshow}{coxph}(obj, nq, years, ...)
}
\arguments{
\item{obj}{model de cox}

\item{nq}{nombre de grups}

\item{years}{anys a calcular el risc observat}

\item{...}{further arguments}
}
\description{
HosmlerLemeshow.coxph
Realitza el test de Hosmer-Lemeshow a partir d'un model de Cox
}
