% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HosmerLemeshow.r
\name{HosmerLemeshow.default}
\alias{HosmerLemeshow.default}
\title{HosmlerLemeshow.glm
Realitza el test de Hosmer-Lemeshow "clàssic"}
\usage{
\method{HosmerLemeshow}{default}(obj, y, nq = 10)
}
\arguments{
\item{obj}{risk estimat o variable predictora}

\item{y}{variable binària (0/1)}

\item{nq}{nombre de grups (per defecte 10)}
}
\description{
HosmlerLemeshow.glm
Realitza el test de Hosmer-Lemeshow "clàssic"
}
