% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HosmerLemeshow.r
\name{HosmerLemeshow.glm}
\alias{HosmerLemeshow.glm}
\title{HosmlerLemeshow.glm
Realitza el test de Hosmer-Lemeshow a partir d'un model logístic}
\usage{
\method{HosmerLemeshow}{glm}(obj, ...)
}
\arguments{
\item{obj}{model logístic}

\item{...}{altres arguments de HosmerLemeshow.Surv (com nq, years,...)}
}
\description{
HosmlerLemeshow.glm
Realitza el test de Hosmer-Lemeshow a partir d'un model logístic
}
