% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HosmerLemeshow.r
\name{HosmerLemeshow}
\alias{HosmerLemeshow}
\title{HosmeLemeshow
Calcula el test de HosmerLemeshow per a logística i per a supervivència}
\usage{
HosmerLemeshow(obj, ...)
}
\description{
HosmeLemeshow
Calcula el test de HosmerLemeshow per a logística i per a supervivència
}
