% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idi_surv.R
\name{idi}
\alias{idi}
\title{idi
Calcula l'IDI (índex de reclassificació)}
\usage{
idi(dat, rr1, rr2, outcome)
}
\arguments{
\item{dat}{dades}

\item{rr1}{risc segons el model 'vell'}

\item{rr2}{risc segons el model 'nou'}

\item{outcome}{nom de la variable cas-control (binaria)}
}
\description{
idi
Calcula l'IDI (índex de reclassificació)
}
