% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.ACCESS.r
\name{import.ACCESS}
\alias{import.ACCESS}
\title{import.ACCESS
for internal use only. Use import.ACCESS2}
\usage{
import.ACCESS(file, nom.taula, nom.variables = list("*"),
  nom.taula.R = nom.taula)
}
\description{
import.ACCESS
for internal use only. Use import.ACCESS2
}
