% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.ACCESS2.r
\name{import.ACCESS2}
\alias{import.ACCESS2}
\title{import.ACCESS2
IMPORTA UNA TAULA DE ACCESS I INCORPORA DICCIONARI DE VARIABLES I FA UNA EXPORTACIO A SPSS}
\usage{
import.ACCESS2(file.mbd, nom.taula, nom.variables = list("ALL"),
  nom.dicc = "", file.spss = "", var.dicc = "",
  noms.taules = gsub(" ", "_", nom.taula), fix.formats = TRUE)
}
\arguments{
\item{file.mbd}{nom del fitxer access}

\item{nom.taula}{vector amb el nom de les taules a llegir}

\item{nom.variables}{llista de vectors amb els noms de les variables a llegir de cada taula}

\item{nom.dicc}{nom de la taula diccionari}

\item{file.spss}{nom del fitxer spss si es vol guardar en SPSS}

\item{var.dicc}{vector amb els noms de les variables del diccioanri amb l'ordre noom, etiqueta variable, etiqueta valors, taula}

\item{noms.taules}{vector amb el nom de les taules com es guardaran a R}

\item{fix.formats}{arregla els formats de les variables (default TRUE)}
}
\description{
import.ACCESS2
IMPORTA UNA TAULA DE ACCESS I INCORPORA DICCIONARI DE VARIABLES I FA UNA EXPORTACIO A SPSS
}
