% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.r
\name{intervals.coxph}
\alias{intervals.coxph}
\title{intervals.coxph
Calcula els intervals de confiança d'un model de cox}
\usage{
\method{intervals}{coxph}(x, level = 0.95, fun = exp, lab = "hr")
}
\arguments{
\item{level}{nivell de confiança (default 0.95)}

\item{fun}{funció per a tranformar els coeficients (per defecte l'exponencial -HR-)}

\item{lab}{etiqueta per al coeficient (defect 'hr')}

\item{objecte}{model de cox}
}
\description{
intervals.coxph
Calcula els intervals de confiança d'un model de cox
}
