% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.r
\name{intervals.glm}
\alias{intervals.glm}
\title{intervals.glm
Calcula els intervals de confiança d'un model logístic (glm)}
\usage{
\method{intervals}{glm}(x, level = 0.95, fun = exp, lab = "or")
}
\arguments{
\item{level}{nivell de confiança (default 0.95)}

\item{fun}{funció per a tranformar els coeficients (per defecte l'exponencial -OR-)}

\item{lab}{etiqueta per al coeficient (defect 'or')}

\item{objecte}{model logístic}
}
\description{
intervals.glm
Calcula els intervals de confiança d'un model logístic (glm)
}
