% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.r
\name{intervals}
\alias{intervals}
\title{intervals
Calcula els intervals de confiança d'un model (lineal, logístic o cox)}
\usage{
intervals(object, level = 0.95, ...)
}
\arguments{
\item{level}{nivell de confiança (default 0.95)}

\item{...}{further argumetns.}

\item{objecte}{(model lineal, logístic o cox)}
}
\description{
intervals
Calcula els intervals de confiança d'un model (lineal, logístic o cox)
}
