% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.variables.r
\name{match.variables}
\alias{match.variables}
\title{match.variables
PROGRAMA QUE BUSCA LES VARIABLES DE 'dades1' A 'dades2'.
LES VARIABLES QUE ES DIUEN IGUAL MIRA LA CORRESPOND?NCIA DE VALORS, SEMPRE QUE NO N'HI HAGIN MOLTS DE DIFERENTS (p.e.10).
LA RESTA DE VARIABLES LES BUSCA A 'dades2' A TRAV?S D'UNA CORRESPOND?NCIA NO EXACTE (veure funci? 'agrep').
DE MOMENT L'OUTPUT ?S UN LLISTAT A LA CONSOLA DE R.}
\usage{
match.variables(dades1, dades2, max.values = 10, max.dist = 0.1,
  file.output = "xxx.doc")
}
\arguments{
\item{dades1}{primera base de dades}

\item{dades2}{segona base de dades}

\item{max.values}{nombre de valors diferents per sobre del qual es considera contínua i per sota es considera categòrica}

\item{max.dist}{distancia per a considerar outlier}

\item{file.output}{nom de l'arxiu on es genera l'informe}
}
\description{
match.variables
PROGRAMA QUE BUSCA LES VARIABLES DE 'dades1' A 'dades2'.
LES VARIABLES QUE ES DIUEN IGUAL MIRA LA CORRESPOND?NCIA DE VALORS, SEMPRE QUE NO N'HI HAGIN MOLTS DE DIFERENTS (p.e.10).
LA RESTA DE VARIABLES LES BUSCA A 'dades2' A TRAV?S D'UNA CORRESPOND?NCIA NO EXACTE (veure funci? 'agrep').
DE MOMENT L'OUTPUT ?S UN LLISTAT A LA CONSOLA DE R.
}
