% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.r
\name{matching}
\alias{matching}
\title{matching
Aparallament casos i controls}
\usage{
matching(dades, nom.casecon, valor.cas, nom.id, num.controls, var.match,
  tol, nom.pair, seed.casos = 321, seed.controls = 678)
}
\arguments{
\item{dades:}{nom de la base de dades en R}

\item{nom.casecon:}{nom de la variable que distingeix entre casos i controls}

\item{valor.cas:}{valor identificador de cas en la variable nom.casecon}

\item{nom.id:}{nom de la variable identificadora (?nica per individu)}

\item{num.controls:}{n?mero de controls a matchejar per a cada cas}

\item{var.match:}{vector character de les variables a matchejar}

\item{tol:}{tolerància per a les variables var.match (amb el mateix ordre)
tol=c(0,2)    0, +-2     //  tol=list(c(0,0),c(0,2))     0, [0,2]}

\item{nom.pair:}{nom de la variable parella}

\item{seed.casos:}{llavor per als casos}

\item{seed.controls:}{llavor per als controls}
}
\description{
matching
Aparallament casos i controls
}
\examples{
\dontrun{
ncases=10
ncontrols=ncases*2
x<-data.frame(id=1:ncases,edad=trunc(rnorm(ncases,60,10)),sexe=rbinom(ncases,1,0.3),cascon=rep(1,ncases))
y<-data.frame(id=1:ncontrols+1000,edad=trunc(rnorm(ncontrols,53,10)),sexe=rbinom(ncontrols,1,0.5),cascon=rep(0,ncontrols))
d<-rbind(x,y)
# difer?ncia d'edat entre casos i controls de 5 anys i mateix sexe.
dm<-matching(dades=d,
         nom.casecon="cascon",
         valor.cas=1,
         nom.id="id",
         num.controls=1,
         var.match=c("sexe","edad"),
         tol=c(0,5),
         nom.pair="pair",
         seed.casos=321,
         seed.controls=678)

d<-merge(d,dm,by="id",all.x=TRUE)
# controls m?s grans que els casos per? no m?s de 5 anys i mateix sexe
dm<-matching(dades=d,
         nom.casecon="cascon",
         valor.cas=1,
         nom.id="id",
         num.controls=1,
         var.match=c("sexe","edad"),
         tol=list(c(0,0),c(0,5)),
         nom.pair="pair",
         seed.casos=321,
         seed.controls=678)
d<-merge(d,dm,by="id",all.x=TRUE)
}
}
