% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge2.r
\name{merge2}
\alias{merge2}
\title{merge2
Fa el mateix que la funció 'merge' però no perd les etiquetes}
\usage{
merge2(x, y, by.id, ...)
}
\arguments{
\item{x}{primera base de dades}

\item{y}{segona base de dades}

\item{by.id}{nom de la variable o vector amb el nom de les variables per ajuntar les dues base de dades (veure argument 'by' de 'merge')}

\item{...}{further arguments from 'merge'}
}
\description{
merge2
Fa el mateix que la funció 'merge' però no perd les etiquetes
}
