% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrients.conversion.r
\name{nutrients.conversion}
\alias{nutrients.conversion}
\title{nutrients.conversion
Cálculo de la cantidad en gramos de cada nutriente diariamente consumida por cada individuo}
\usage{
nutrients.conversion(file.alim, file.coef)
}
\arguments{
\item{file.alim}{fitxer SPSS amb els aliments}

\item{file.coef}{fitxer SPSS amb els coeficients per a fer la conversió}
}
\description{
nutrients.conversion
Cálculo de la cantidad en gramos de cada nutriente diariamente consumida por cada individuo
}
