% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order2.r
\name{order2}
\alias{order2}
\title{order2
ORDENA UNA BASE DE DADES (data) SEGONS UNES VARIABLES (vars) SENSE PERDRE ELS ATRIBUTS.}
\usage{
order2(data, vars, vars.view = "ALL", decreasing = FALSE)
}
\arguments{
\item{data}{dades}

\item{vars}{vector de les variables que serviran per ordenar les files}

\item{vars.view}{vector de les variables que es guardaran de la base de dades (default is 'ALL')}

\item{decreasing}{default is FALSE}
}
\description{
order2
ORDENA UNA BASE DE DADES (data) SEGONS UNES VARIABLES (vars) SENSE PERDRE ELS ATRIBUTS.
}
\note{
no funciona si es vol ordenar per a més d'una variable
}
