% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HosmerLemeshow.r
\name{plot.HosmerLemeshow}
\alias{plot.HosmerLemeshow}
\title{plot.HosmerLemeshow
mètode genèric per graficar el resultat de Hosmer Lemeshow}
\usage{
plot.HosmerLemeshow(hl, perc = FALSE, leg.text = c("Expected",
  "observed"), fact = 1, ...)
}
\arguments{
\item{hl}{objecte de classe HosmerLemeshow}

\item{perc}{printa el percentatge o el valor absolut d'esdeveniments}

\item{leg.text}{text}

\item{fact}{factor per multiplicar els esdeveniments}

\item{...}{further arguments}
}
\description{
plot.HosmerLemeshow
mètode genèric per graficar el resultat de Hosmer Lemeshow
}
