% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potencia.r
\name{potencia}
\alias{potencia}
\title{potencia
calcula la m?nima diferencia detectada per a una potencia donada i un nivell de confian?a donat
tamb? dibuixa un plot de la potencia a un nivell de confian?a donat}
\usage{
potencia(SE, pot = 0.8, alpha = 0.05, df = Inf, digits = 2,
  fact = 1, max.rang = 5, OR = TRUE, plot = TRUE,
  max.diff.est = 10)
}
\arguments{
\item{SE}{error est?ndar de la beta}

\item{pot}{pot?ncia (per defecte al 80%)}

\item{alpha}{nivell de significaci? (per defecte 5%)}

\item{df}{graus de llibertat (per estudis amb grand?ries de mostra molt petits es posa la grand?ria de la mostra, si l'estudi ?s gran, es posa Inf).}

\item{digits}{nombre de decimals que vols que aparegui en el gr?fic}

\item{fact}{vols multiplicar la variable predictora per algun valor? ?til per si es vol calcular la pot?ncia per al augment de 10 unitats de colesterol total per exemple, es posaria fact=10.}

\item{max.rang}{l?mits del gr?fic (+/-max.rang*SE)}

\item{OR}{vols donar els restulats en escala d'Odds Ratio ? Hazard Ratio (exp(beta)) o no (beta).}

\item{plot}{vols que es dibuixi el gr?fic o nom?s que surtin els resultats}

\item{max.diff.est}{la m?nima difer?ncia a detectar es buscar? entre 0 i SE*max.diff.est. Per defecte ?s 10, per? si el c?lcul d?na errors de converg?ncia es pot modificar.}
}
\description{
potencia
calcula la m?nima diferencia detectada per a una potencia donada i un nivell de confian?a donat
tamb? dibuixa un plot de la potencia a un nivell de confian?a donat
}
\examples{
\donttest{
# Exemple 1: Regressi? lineal
#  pot?ncia 80\%
#  alpha: 5\%
#  SE de la beta: 0.5
potencia(0.5,pot=0.8,alpha=0.05,OR=FALSE)
# -> la difer?ncia m?nima a detectar amb una pot?ncia del 80\% i un nivell de significaci? del 5\% (i amb un error est?ndar de la beta), ?s de 1.4
# Exemple 2: Odds Ratio
# pot?ncia 80\%
# alpha: 5\%
# regressi? lineal -> OR=FALSE
# SE de la beta (log(OR)): 0.1
potencia(0.1,pot=0.8,alpha=0.05,OR=TRUE)
#-> l'OR m?s petit que es pot detectar amb una pot?ncia del 80\% i un nivell de significaci? del 5\% (i amb un error est?ndar de la beta), ?s de 1.32 (si ?s de risc) i de 0.76 (si ?s protector).
}
}
