% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.r
\name{prepare}
\alias{prepare}
\title{prepare
prepare data: labels variables from vari.label attribut, and convert factors those variables with value.labels atributes.}
\usage{
prepare(dat, expand = FALSE)
}
\arguments{
\item{dat}{dades}

\item{expand}{codifica valors sense etiquetes de valor (per exemple 1=diab i 2=no diabétic i també apareixen 9's, aleshores el 9 serà una nova categoria, si no el 9 passarà a ser NA).}
}
\description{
prepare
prepare data: labels variables from vari.label attribut, and convert factors those variables with value.labels atributes.
}
