% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HosmerLemeshow.r
\name{print.HosmerLemeshow}
\alias{print.HosmerLemeshow}
\title{print.HosmerLemeshow
mètode genèric per printar el resultat de Hosmer Lemeshow}
\usage{
\method{print}{HosmerLemeshow}(hl)
}
\arguments{
\item{hl}{objecte de classe HosmerLemeshow}
}
\description{
print.HosmerLemeshow
mètode genèric per printar el resultat de Hosmer Lemeshow
}
