% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.r
\name{print.intervals}
\alias{print.intervals}
\title{print.intervals
métode per a printar per consola}
\usage{
\method{print}{intervals}(n, len = 6, d = 3,
  exclude.intercept = inherits(n, "intervals.glm"))
}
\arguments{
\item{n}{objecte intervals}

\item{len}{length (default 6)}

\item{d}{nombre de digits (default 3)}

\item{exclude.intercept}{logical. incloure intercept (no si el model és cox)}
}
\description{
print.intervals
métode per a printar per consola
}
