% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print2.r
\name{print2}
\alias{print2}
\title{print2
FUNCIO QUE VISUALITZA (a trav?s de la funci? 'fix') UNA MATRIU O UN DATA.FRAME
té com a opcio que mostri les etiquetes de valors enlloc dels valors.}
\usage{
print2(data, view = 1)
}
\arguments{
\item{data}{dades}

\item{#}{view =1: noms etiquetes, view=2: valor i etiqueta, view!=1 ? 2: noms valor}
}
\description{
print2
FUNCIO QUE VISUALITZA (a trav?s de la funci? 'fix') UNA MATRIU O UN DATA.FRAME
té com a opcio que mostri les etiquetes de valors enlloc dels valors.
}
