% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spss4.r
\name{read.spss4}
\alias{read.spss4}
\title{read.spss4
reads SPSS files detecting automatically date variables and transforming them to chron format}
\usage{
read.spss4(file, elimina.miss = TRUE, tolower = TRUE,
  use.value.labels = FALSE, keep.var = "ALL",
  date.format = "d-mon-Y", reencode = NA, ...)
}
\arguments{
\item{file}{.sav file}

\item{elimina.miss}{elimina els missing definits per l'usuari en el fitxer SPSS (default TRUE)}

\item{tolower}{converteix els noms a minúscules (default TRUE)}

\item{use.value.labels}{converteix els codis a etiquetes (similar a read.spss de foreign)}

\item{keep.var}{vector de les variables a llegir (default 'ALL')}

\item{date.format}{format de les dates (veure 'out.format' de la funció 'chron')}

\item{reencode}{codi UTF-8, latin1, ...}

\item{...}{further arguments passed to `read.spss`}
}
\description{
read.spss4
reads SPSS files detecting automatically date variables and transforming them to chron format
}
