% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.plot.r
\name{ROC.plot}
\alias{ROC.plot}
\title{ROC.plot
DIBUIXA UNA CURVA ROC I DONA L'AUC I EL SEU INTERVAL DE CONFIANÇA}
\usage{
ROC.plot(y, p, alpha = 0.05)
}
\arguments{
\item{y}{variable binaria 0/1}

\item{p}{valors predits}

\item{alpha}{nivell de significació (default 0.05)}
}
\description{
ROC.plot
DIBUIXA UNA CURVA ROC I DONA L'AUC I EL SEU INTERVAL DE CONFIANÇA
}
