% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_varlist.r
\name{spss_varlist}
\alias{spss_varlist}
\title{spss_varlist
Funci贸 que captura la meta-informaci贸 d'un fitxer d'SPSS (nom de les variables, etiquetes, formats, ...)}
\usage{
spss_varlist(file)
}
\arguments{
\item{files}{fitxer SPSS .sav}
}
\value{
Returns an R data.frame object with a row for each variable, columns in this
frame are:
\itemize{
\item varname		Variable Name, 8-char upper case.
\item varlabel		Variable label, up to 40 char in length.
\item printfmt		Spss PRINT FORMAT mnemonic (A40, F8.2, DATE11, etc).
\item nmiss		An integer value indicating how many missing values
                      are defined, 0, 1, 2, or 3. Can also take values -2 or
                      -3 indicating that  Missing Value 1 and Missing Value 2
                      represent a range.
\item missing1		First missing value (or NA).
\item missing2		Second missing value (or NA).
\item missing3		Third missing value (or NA).
\item longname		Long variable name (SPSS version >= 12)
}
}
\description{
spss_varlist
Funci贸 que captura la meta-informaci贸 d'un fitxer d'SPSS (nom de les variables, etiquetes, formats, ...)
}
\details{
spss_varlist(FILE) -- extract info about variables from header of
                       FILE, an SPSS for Windows data file (.sav)

Based on the Perl script  'spssread.pl'  :

 "spssread.pl - a utility to print info about SPSS data files
 version 0.2.1 on 16 Jun 2005 by Scott Czepiel <sczepiel@gmail.com>
 for updates see: <http://czep.net/data/spssread/>"

Format code mnemonics, and Record Type layout information, obtained from
a document (presumably authored by SPSS Inc.) made available
at <http://www.wotsit.org>, a site containing file format
information on hundreds of different file types.

Adapted for use in R by Dave Macfarlane and Isaac Subirana.
April 2006.
(dave@imim.es, isubirana@imim.es)
}
\note{
A partir de la versi? 12.0 (inclosa) apareix en l'arxiu els longnames: tants com variables hi ha i que no tenen perqu? coincidir amb els varnames.
per a versions anteriors cal ignorar aquesta informaci? (que apareix al final de l'arxiu) ja que pot ser que no es corresponguin
els longnames amb els varnames ni en n?mero (aix? passa quan inicialment s'ha creat un arxiu .sav amb una versio >=12 i posteriorment
}
\author{
Dave McFarlane, Isaac Subirana
}
