% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset2.r
\name{subset2}
\alias{subset2}
\title{subset2
ANÀLEG AL SUBSET PER NO PERD ELS ATRIBUTS}
\usage{
subset2(data, keep, vars, ...)
}
\arguments{
\item{data}{dades}

\item{keep}{logical vector per seleccionar les files. O una expressió lógica en format caràcter}

\item{vars}{vector amb els noms de les variables a seleccionar}
}
\description{
subset2
ANÀLEG AL SUBSET PER NO PERD ELS ATRIBUTS
}
\note{
A DIFER?NCIA DEL SUBSET, ES NECESSARI PASSAR-LI UN DATA.FRAME O UNA MATRIU.
TAMBE SE LI HI DE POSAR EL KEEP COM UNA EXPRESSI EN FORMAT CADENA, I AIXI NO CAL POSAR EL DATAFRAME.
}
