% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.r
\name{table2}
\alias{table2}
\title{table2
funcion que modifica el \code{table} que substitueix els valors de les variables per l'etiqueta i el valor; per tant surten les etiquetes a la taula.}
\usage{
table2(..., max.values = Inf, margin = NULL, digits = NULL,
  print = TRUE)
}
\arguments{
\item{...}{vector o vectors pel qual es farà la taula de freqüències}

\item{max.values}{nombre màxim de categories (default Inf - vol dir que no hi ha límit)}

\item{margin}{1 proporció per files, 2 proporció per columnes, 3 per capes, etc. (NULL no calcula proporcions sinó freq absolutes)}

\item{digits}{nombre de decimals}

\item{print}{printa per la consola (default TRUE)}
}
\description{
table2
funcion que modifica el \code{table} que substitueix els valors de les variables per l'etiqueta i el valor; per tant surten les etiquetes a la taula.
}
\note{
els arguments de table2 han de ser vectors (o variables d'un data.frame) de la mateixa longitud.
}
